backend/backend/services/auth_service.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
import bcrypt
import os

# 🔐 Secret key (use env in production)
SECRET_KEY = os.getenv("SECRET_KEY", "supersecretkey")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24


# ✅ Hash password
def hash_password(password: str) -> str:
    # Fix: bcrypt max 72 bytes
    password = password[:72]

    hashed = bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt())
    return hashed.decode("utf-8")


# ✅ Verify password
def verify_password(password: str, hashed_password: str) -> bool:
    password = password[:72]

    return bcrypt.checkpw(
        password.encode("utf-8"),
        hashed_password.encode("utf-8")
    )


# ✅ Create JWT token
def create_access_token(data: str):
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)

    payload = {
        "sub": data,
        "exp": expire
    }

    token = jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
    return token


# ✅ Decode JWT token
def decode_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload.get("sub")
    except JWTError:
        return None