backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
scikit-learn
email-validator

motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]
passlib[bcrypt]
openai
reportlab==4.4.10
bcrypt